emotion: try to correctly fix the fill value.
SVN revision: 64255

+++ legacy/emotion/src/modules/gstreamer/emotion_gstreamer.h
@@ -132,8 +132,8 @@ struct _Emotion_Gstreamer_Video
 
    struct
    {
-      int            width;
-      int            height;
+      double         width;
+      double         height;
    } fill;
 
    Eina_Bool         play         : 1;
